Add WIP activity schema plus tests. Run 'npm install'
Tests depend on JSV node package.

The schema is a work in progress. (a) it needs to be further specified, 
as there are probably many invalid json objects that would currently be 
marked as  "valid" by this, and (b) it should be added to as we create 
new runtime examples