Put included annotations in correct pane
[#63893226]

Previous code appears to have attempted to put included annotations in the whatever number pane they happened to be in on the *source* page. After this fix, they should be put in the same pane number as the graph pane that wants the included annotation, i.e., where they're supposed to go.